+++ src/Constant/ContentType.php
@@ -1,4 +1,4 @@
-﻿<?php
+<?php
 /*
  * Licensed to the Apache Software Foundation (ASF) under one
  * or more contributor license agreements.  See the NOTICE file
@@ -18,35 +18,20 @@
  * under the License.
  */
 
-
-class MessageDigestUtil
+namespace Aliyun\ApiGatewaySign\Constant;
+/**
+ * 常用HTTP Content-Type常量
+ */
+class ContentType
 {
-	/**
-    * md5和base64处理
-    *
-    * @param $input
-    * @return
-    */
-	public static function Base64AndMD5($input) 
-	{
-		if ($input == null || strlen($input) == 0) {
-			throw new Exception("input can not be null");
-		}
-
-		return base64_encode(md5(unpack('C*', $input)));
-	}
-
-	/**
-    * UTF-8编码转换为ISO-9959-1
-    *
-    * @param str
-    * @return
-    */
-    public static function Utf8ToIso88591($input)
-    {
-		if ($input == null || strlen($input) == 0) {
-			return $input;
-		}
-		return mb_convert_encoding($input, "ISO-8859-1", "UTF-8");
-    }
+    //表单类型Content-Type
+    const CONTENT_TYPE_FORM = "application/x-www-form-urlencoded; charset=UTF-8";
+    // 流类型Content-Type
+    const CONTENT_TYPE_STREAM = "application/octet-stream; charset=UTF-8";
+    //JSON类型Content-Type
+    const CONTENT_TYPE_JSON = "application/json; charset=UTF-8";
+    //XML类型Content-Type
+    const CONTENT_TYPE_XML = "application/xml; charset=UTF-8";
+    //文本类型Content-Type
+    const CONTENT_TYPE_TEXT = "application/text; charset=UTF-8";
 }
