+++ src/Util/MessageDigestUtil.php
@@ -1,4 +1,4 @@
-<?php
+﻿<?php
 /*
  * Licensed to the Apache Software Foundation (ASF) under one
  * or more contributor license agreements.  See the NOTICE file
@@ -17,32 +17,40 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-spl_autoload_register("Autoloader::autoload");
 
-class Autoloader
-{	
-	private static $autoloadPathArray = array(
-		"Constant",
-		"Http",
-		"Util"
-	);
-	
-	public static function autoload($className)
-	{
-		foreach (self::$autoloadPathArray as $path) {
-			$file = dirname(__DIR__).DIRECTORY_SEPARATOR.$path.DIRECTORY_SEPARATOR.$className.".php";
-			$file = str_replace('\\', DIRECTORY_SEPARATOR, $file);
-			if(is_file($file)){
-				include_once $file;
-				break;
-			}
-		}
-	}
-	
-	public static function addAutoloadPath($path)
-	{
-		array_push(self::$autoloadPathArray, $path);
-	}
+
+namespace Aliyun\ApiGatewaySign\Util;
+
+use Exception;
+
+class MessageDigestUtil
+{
+    /**
+     * md5和base64处理
+     *
+     * @param $input
+     * @return
+     */
+    public static function Base64AndMD5($input)
+    {
+        if ($input == null || strlen($input) == 0) {
+            throw new Exception("input can not be null");
+        }
+
+        return base64_encode(md5(unpack('C*', $input)));
+    }
+
+    /**
+     * UTF-8编码转换为ISO-9959-1
+     *
+     * @param str
+     * @return
+     */
+    public static function Utf8ToIso88591($input)
+    {
+        if ($input == null || strlen($input) == 0) {
+            return $input;
+        }
+        return mb_convert_encoding($input, "ISO-8859-1", "UTF-8");
+    }
 }
-
-?>
